IJPL-177555 Do not consider dropdown elements for severity preferred size
If the preferred size is larger than the cell size, the full size component
appears on mouse hover. However, when the current value is short enough,
it doesn't bring any value and looks strange.

By turning off using dropdown values for preferred size calculation
we revert to the default behavior, which only shows the hover popup
if the current value is too large.

GitOrigin-RevId: f1e1e867e2a4ec2eee1622d3351f765972e82c70