IJPL-177555 Use separate components for highlighting renderers and editors
The same fix as for the severity column.

The renderer is reused over and over again to render cells,
even while the editor is shown. The editor, on the other hand,
exists in one place while the editing is being performed.
Using one component for both creates visual glitches,
as it can't be in two places at the same time, acting as the active
editor and a temporary renderer elsewhere.

Fix by creating two separate instances.

GitOrigin-RevId: b164978380f20db5d62b74be4e953d9dc8a8b424