IJPL-177555 Do not consider dropdown elements for highlighting preferred size
The same fix as for the severity renderer.

If the preferred size is larger than the cell size, the full size component
appears on mouse hover. However, when the current value is short enough,
it doesn't bring any value and looks strange.

By turning off using dropdown values for preferred size calculation
we revert to the default behavior, which only shows the hover popup
if the current value is too large.

GitOrigin-RevId: f7d1b483e7e48c661148dac28bb8845f8222795d