IJPL-177555 Make ComboBoxTableRenderer close on second click
We already have a generic mechanism for this.
StackingPopupDispatcherImpl will prevent mouse event processing
if the mouse click was within the bounds of the component
that shown the popup to begin with.
We only need to explicitly specify which component is that.
No leaks there either, as this thing uses weak references
internally.

GitOrigin-RevId: d2716ae6fabb4e5bd458c985acf7d6665995c974