[rd] IJPL-181961 Get rid of `intellij.platform.ide.core` in `intellij.platform.kernel`
We want to have a dependency on kernel from `ide.core` to propagate context to `launchOnShow`.
Right now it's not possible because it'll cause circular dependency.

The only reason to have a dependency on `ide.core` inside kernel is using `ApplicationInitializedListener`.
I'm moving `KernelApplicationInitializedListener` to a separate `kernel.impl` module to get rid of this dependency.

GitOrigin-RevId: b6d21b4446616d2c3d576da5520a99132cb892f8