IJPL-164422 DocumentImpl: do not create new changedPart if it's a String
This code was initially written to avoid dragging along with
the changed part some huge chunks of text in case this value
is some complicated implementation of CharSequence.
But if it's just a String, turning it into a smart immutable char sequence
has no benefit but may slow down things.

Notably, this change noticeably improves the performance
of LineSet.isSingleLineChange.

GitOrigin-RevId: 86ef29d3ba08dc7eaf7a51127449afcb92ae6d27