[terminal] IJPL-182160 Do not update terminal output PSI on every document change
Previously, PsiDocumentManager scheduled the document committing on every change. It led to very frequent write actions when the terminal received a lot of output, and therefore performance problems (for example, main editor typing was sluggish).

The solution is to implement custom LightVirtualFile that will disable the event system, so PsiDocumentManager won't try to commit the document on its own.
But then we need to commit it manually at some time. Do it with 50ms debouncing in the `updatePsiOnOutputModelChange` function.

GitOrigin-RevId: 9d1ae31206b8ca824e44951cc0d474fe0ee945c7