PY-78919, KTNB-954: Fix color scheme change
Use `DelegateColorScheme` because it's the only way to update it (see `EditorImpl.MyColorSchemeDelegate#updateGlobalScheme`). Also, `DelegateColorScheme` isn't unwrapped in `setColorsScheme`; only `MyColorSchemeDelegate` is unwrapped there, so the comment was adjusted to reflect it.

To solve a chicken-and-the-egg problem, unwrapping is now performed in `DefaultNotebookEditorAppearance`: we need non-delegate there because a delegated scheme may contain old color value.


Merge-request: IJ-MR-158588
Merged-by: Ilya Muradyan <Ilya.Muradyan@jetbrains.com>

GitOrigin-RevId: 78b14efd086af552f69044bd8d2bca2e6ea32e47