[platform] wait until ApplicationInfo is loaded before setting the frame class in XToolkit (IJPL-165691)
Otherwise, when an embedded variant of the frontend process (JetBrains Client) is started, it may not load ApplicationInfo, and the frame class won't be set leading to an incorrect icon in the taskbar.

Also, if an attempt to set the frame class fails with an exception, a warning should be printed to the log.

GitOrigin-RevId: d47ea5ba0d0dd1b14b2a75d278490d4dd6007cce