ML in SE: Introduce SearchEverywhereTab as a single source of truth (IJPL-181860)
To prevent issues like IJPL-181860, where the behavior of the tab
is defined in multiple places, this commit introduces a single
instance of every supported Search Everywhere Tab that exposes
the required methods for determining the behavior of that tab.

There is no synchronization between experiments and settings,
so it may happen that "disabled ML ranking" is in fact enabled,
through an experiment.

The only case where the ML ranking is not enabled through
an experiment is when ML ranking was enabled by default,
but it was disabled by the user.

GitOrigin-RevId: cb18038497e8ca9e6d4a62d8256913f7c5b97d98