[wsl][IJPL-181144] introduced `JdkUtil#isCompatible(Path, Project)` and `checkForJdk(Path, Boolean)` marked as obsolete
`checkForJdk(Path, Boolean)` depends on a flag with information about the current execution environment. As a result, this flag could be miss-used and the wrong check would be performed. The method should only check "is it possible to use a JDK instance" no matter to which execution environment it belongs to.
`JdkUtil#isCompatible` is a replacement for environment-related logic from `checkForJdk(Path, Boolean)`.

GitOrigin-RevId: d6e9346c669cf2ad95158867c7ce58e51082b4d5