IJPL-177087 [rd][debugger] Deduplicate execution stacks and stack frames, part one: deduplicate IDs to ensure equality
Both on backend and frontend parts. IDEA's debugger engine relies on equality in many parts, so we want to be sure that instances that are equal in the old monolith remain equal in the frontend. For that, we ust ensure that two stacks/frames that are same by equals have the same ID.

Some parts of the UI also require that instances be the same by object identity (== in Java, === in Kotlin), that's what part two addresses.

GitOrigin-RevId: 6735cec276ba7d64672698933b0e8b7929dadf1a