[fleet] introduce fleet.dock.runtime module layer
Now we have 3 module layers in dock: runtime, api and desktop. Runtime is encapsulated all the dependencies which are not dock API but exposed to the entire application. Logging API and Window API are moved there as it's the only way to properly load java services implementations: KLoggerFactory and MainDispatcherFactory respectively.

GitOrigin-RevId: 888cd0f50433b984fc52c441607c554ac8f9687e