[JEWEL-792] Add slot-based API for InlineBanner
This adds a slot-based API to InlineBanner, in addition to
the String-based API that already exists. It is a
non-breaking change. The reason is to allow users to specify
their own main content, e.g., to use Markdown. The title is
unaffected as it remains plain text.

Additionally, this fixes the KDoc for InlineBanner by making
sure there is no content after the tags section. Before this
all the content after the last tag would be considered to be
part of the tag itself's documentation, which is incorrect.

The KDoc have also been rewritten for correctness and to
improve clarity, making them simpler and more readable.