[terminal] IJPL-182693 Add extension point: `CUSTOM_LOOKUP_EP` for custom `TerminalLookup` creation
This is a different approach that avoids using the previous putLookup function. It provides a less "rude" way to insert a custom lookup. I added `TerminalInput.KEY` to store the `terminalInput` in the `editor`, so it can be used when creating the `TerminalLookup`.


(cherry picked from commit b25f60caf3f12474076c116422b8e37b327663d7)

IJ-MR-157771

GitOrigin-RevId: 6aed350ed3e757f51a8152cbcf9e5a839c414411