IJPL-177087 [rd][debugger] Send initial XStackFrame presentation with DTO, not acquire it later via RPC
There is a problem that XFramesList relies on customizePresentation being calculated synchronously, which doesn't work well with the RPC approach. And it's not really required to do customizePresentation in a flow (as it's synchronous anyway and is called only once).

Here, the initial presentation is passed with a DTO, and is never being recalculated since. There's a risk that some XStackFrame implementations might change presentation over time, and right now we wouldn't be aware of that (but TBH, neither would we with the previous approach with RPC and a flow). If that's the case, then we'll think it further.

GitOrigin-RevId: 372a6345eb5e2ea0dac42844d5de8dc56f714d96