PY-80128: Update bundled venv because an old one is incompatible with the latest.
We use bundled venv to create virtual envs because system python might not have one. However, if it does, it must be compatible with the one we bundle. A new version was released 31.mar.2025. When installed, it breaks our bundled one.

So, we bundled a version. It doesn't work with Py2 though, dus we now check language level.

GitOrigin-RevId: cc7ff994bf32ef4333908321c6bb189a0e9ca9e5