Python tests: store a path to conda in `PyCondaPackageService`.
While one should use a conda path from `PyCondaFlavorData`, some legacy code still depends on this service.

This service does a good job trying to find conda on a user PC, but obviously fails in test as conda is installed in a usual location.

We store a conda path in this service to make tests work

GitOrigin-RevId: d8e1d08b268ae5259e7050c5c9bb8cc8184e3751