PY-80143 Move BALLOON_NOTIFICATIONS from a PyStubPackagesAdvertiser static field to a Visitor getter to avoid NPE when calling `getNotificationGroup("Python Stub Packages Advertiser")`
It might be the case that the NotificationGroupManagerImpl.registeredGroups is not initialized when PyStubPackagesAdvertiser is first initialized. For that reason, BALLOON_NOTIFICATIONS calls `getNotificationGroup` each time it's accessed. The call is a hashmap lookup and so is very cheap.

As a bonus, `NotificationGroupManager` initialization is moved from the companion object, so it's not initialized when the PyStubPackagesAdvertiser class is. So no `forceNonCancellableSectionInClassInitializer` is required

GitOrigin-RevId: 4e7f54f1930e38471218828abb4af9a0ee679bb0