[terminal] IJPL-182032 Reset port forwarding watcher delay on line break read TTY connector
`PortForwardingProcessWatcher` checks ports periodically with exponentially increasing delay, so we need to periodically reset this delay when users perform some activity in the terminal.
In the Classic Terminal, there was a heuristic: we reset the delay when we read line break from the TtyConnector. So, it triggers exactly when user submits the command.
Reuse this heuristic in the Reworked Terminal.

GitOrigin-RevId: 736bf0fc943d149f55f7f6591b55f6ad61dea309