[runtime repository builder] add 'intellij.platform.runtime.repository' as a library while building runtime module repository (IJPL-109)
Before, 'intellij.platform.runtime.repository' was added to the build process from the IDE installation. This will cause problems when format of the runtime module repository is changed, because an older version of the IDE may be used for compilation. So it's better to add it as the part of 'devkit.runtime.module.repository.jps' library which is added to the JPS build process' classpath using a fixed version specified in the current project source code.

GitOrigin-RevId: b1c23c3416f778eb746dea4137dd6c51a6ae3b38