[terminal] IJPL-182032 Keep the cursor visible on the screen when the port forwarding panel appears
Previously, when the port forwarding panel appeared on top, the editor was scrolled down, so the cursor position was hidden.
Let's react on port forwarding panel size changes and update the scroll position.
Move top panel management from `ReworkedTerminalWidget` to the `ReworkedTerminalView`. Refactor the TerminalPanel to extract the layered pane logic to the separate `TerminalLayeredPane`.

GitOrigin-RevId: 78992a106db33e5861d362d5aca00942dbf0382c