[TrustedProject] fix: automatically trust projects on in the createProject and newProject functions
The createProject and newProject functions create a new instance for the IDE-generated projects (tests or new project wizards). We completely control the project file generation process and can guarantee that we don't add malicious code.

However, the prepareNewProject function can be executed for any project. This function has a different "new project" semantic with the newProject function. Here "new project" means initializing a new project instance without the project caches (.idea). Therefore, it can be used for the existing project without (or with broken) project caches.

### Issues
  * IJPL-183135 `Preview in safe mode` opens in normal mode

GitOrigin-RevId: baf2a12490791dca716801686ba37eb983cc943a