Revert "IJPL-182457 use IO dispatcher for indexing, instead of Default, and avoid `yield`"
This reverts commit 2e686bcd4f33703b3b9d3c1b12c69adad68cbf1c.

reason: now we need to jump from IO to Default to get RA, and then back. That is, if earlier we went through the list of files, took RA for each and calculated something in it. And all this happened in the same thread. Now we go through the list of files, switch the dispatcher (and, most likely, the thread too), take RA, calculate something in it, then return to the IO dispatcher (which, most likely, also leads to a change of thread), and take the next file. Since IO and Default dispatchers are underloaded, there is a lot of parking/unparking of threads. I'll return it as it was for now.

GitOrigin-RevId: 034178da8e1820de20377e0b14532fe45a10138e