[terminal] IJPL-182500 Make 'Open in Terminal' action work for Reworked Terminal in RemDev
Original `RevealFileInTerminalAction` is executed on backend, because Classic Terminal can open the tab with specified working directory only from the backend side.
While Reworked Terminal now can be opened only from the frontend side.
So, the solution is to create a copy of the action for the Reworked Terminal that will be located on the frontend. Only one action will be enabled simultaneously.

GitOrigin-RevId: 5e17b9225158320552ed3491cafa7ff3f910d0f3