[path-annotations] Add path annotation checks for Path.of() and Path.resolve() usage
Introduced inspections to warn about non-annotated strings in `Path.of()` and `Path.resolve()` contexts. Updated tests to cover these scenarios and added `@NativePath` annotation to relevant parameters in the Docker CLI functionality.

Generated by Junie to highlight `Path.of()` calls with not annotated strings as normal warnings and add highlighting of `Path.resolve()` with not annotated strings as weak warnings.

GitOrigin-RevId: fb55c1716e3c007b35428320f58c5cf7afce5b5f