[java-psi] Simplify PsiPackageImpl.isValid
It's probably ok to consider a package as valid, even if it doesn't contain any directories, and the prefix does not exist. Usually, invalid PsiElement is something deleted, or replaced, which cannot be used anymore. This is likely not the case here.
May help IDEA-368981 Renaming package never finishes, stuck "Looking for Usages"

GitOrigin-RevId: 27a08684daed8fd9232dbab3815e19deddd79b0d