[terminal] IJPL-183387 Fix race of terminal settings synchronization after migration from registry
If you start the new IDE version of the IDE (251+, after updating from 243) for the first time in RemDev mode, migration can be performed simultaneously on frontend and backend. So we can end up with terminal engine being out of sync.
Fix it by running the migration logic only on Frontend. So, the updated terminal engine will be set on the backend after sync. Also, trigger settings save manually because otherwise it will be synced only some time later.

GitOrigin-RevId: 253c2c3e1dd35bdcc58eb21a5ae572d415bccb2b