[terminal] IJPL-182482 Ignore read time during serialization
We need this `readTime` only locally on the backend and put it to the event object only for simplicity.
It is not a good design, but for now there is no suitable alternative. Ideally, we need to have the separate serializable DTO classes hierarchy for RPC.
Or switch to plain longs for measuring the time.

(cherry picked from commit f8e20acdfe0d856c73750711d1a14917fc9b5973)

IJ-CR-160384

GitOrigin-RevId: e12a5a402f4065595b9ed2f25c79bdfb7f5a6734