[terminal] IJPL-182482 Ignore read time during serialization
We need this `readTime` only locally on the backend and put it to the event object only for simplicity.
It is not a good design, but for now there is no suitable alternative. Ideally, we need to have the separate serializable DTO classes hierarchy for RPC.
Or switch to plain longs for measuring the time.

GitOrigin-RevId: f8e20acdfe0d856c73750711d1a14917fc9b5973