[terminal] IJPL-182482 Introduce TimedKeyEvent and use it to pass the information about monotonic time of event init
It will be used to measure the typing latency.
Also, I had to copy the `TerminalEventsHandler` interface to the frontend, because it was necessary to change the parameter types. Previously, we used the interface located in the shared part and common for Gen1 and Gen2.

(cherry picked from commit 6b8b0cccde598dc6cef9217232eb3c6725973f9a)

IJ-CR-160384

GitOrigin-RevId: 60fc41f442152d325248294f0815e188a2187203