[terminal] IJPL-182482 Reimplement backend typing latency measurement
Pass event time and id to the TerminalStarter. Implement the separate method for that in our override for that.
This way is more explicit and safe because it is not using the global state and matching by event object.

(cherry picked from commit ddbf4373bf6226c72bd8db3bc0d8b849d4ba3112)

IJ-CR-160384

GitOrigin-RevId: d737dc7caf38dd4ed61a9c2f0eda7385226345f6