Eel test framework: Eel-parameterized tests introduction.
You are welcome to see:
`com.intellij.platform.testFramework.junit5.eel.params.api` and
and run `EelParametrizedTest` locally.

We also tried to fix the docker-client IOCP thread leak on Windows.
We now close the fixture (disconnect from the Docker) and explicitly
whitelist IOCP on Windows (disconnection might help on other OSes, but
not on Windows).

GitOrigin-RevId: 462ac0d2ba44a8f17e45cbcd1041437d916bccca