[collab] Add some much needed tests for associateCachingBy and fix order
Ordering was broken when removing and adding items in the same update.
Previously, existing items had a priority over new items, causing the list order to desync.

Now, all items of the new list are added to the result map first.
Afterwards, items that were not copied over are removed.

GitOrigin-RevId: dbf321becc7b5267a857812139896d482c0c930e