PY-79488 Try detecting project model roots containing all dependent modules
It's primarily necessary for Poetry projects lacking the declarative structure
of a uv workspace, where there could be a bunch of projects referring to each
other under a common root that itself doesn't have pyproject.toml.

In particular, it's the case with grazie-ml where there are multiple projects
under the "libs/" dir, and neither the root of the project nor "libs/" contains
its own pyproject.toml.

GitOrigin-RevId: 596e90cb5dedd47163f1d018f959fb043eba39ff