[platform tests] temporarily disable the "Do not add dependency on a parent plugin" check in PluginModelValidator
It's not an error to have such a dependency. Currently, have several violations in plugins/evaluation-plugin/ultimate/resources/META-INF/eval-ultimate.xml, for example. When content modules are registered in an additional file included via 'xi:include', removing dependency on the parent plugin will lead to red code in the editor. It's better to include such content modules directly in the root plugin.xml (we now allow referencing content modules from ultimate part of the repository in plugin.xml from intellij-community). But until this is done, it's better to disable the check.

GitOrigin-RevId: 3435c35335239fbbb83549a2d1672d778bc96b8c