[terminal] IJPL-182466 Refactor: extract terminal session creation logic to the separate service
It is worth having a separate place for terminal sessions state, that is not related to the `TerminalTabsManager`. `TerminalTabsManager` is only the client that starts the terminal session to display it in the terminal tool window.
While terminal sessions can be used in other places as well.
Also, it is worth encapsulating the `newValueEntity` and `findValueEntity` API.

GitOrigin-RevId: f446bb6ea681ef6379166c2a5935a66b26d6541c