IJPL-183523, IJPL-184075, IJPL-184079 Terminal font size changes
Introduce ChangeEditorFontSizeStrategy that allows overriding
the editor's behavior when it comes to mouse wheel zoom
and font reset actions.

Use this new API to override the behavior for the terminal editors
and delegate to the new TerminalFontSizeProviderImpl service
to apply the changes.

Make the terminal editors listen to font changes and react.

Note that the editor's change font size actions are not covered
by this new API because the terminal has its own actions for that
(to allow keymap customization) and those should be fixed instead.

(cherry picked from commit 3c4173128df790708b74e8cde5516ce2d6aa3ae1)

IJ-CR-160631

GitOrigin-RevId: 7ef9180d4980c4c31aaf74d0b39b8a69ae03ebad