IJPL-184075 Fix increase/decrease font size action for the reworked terminal
Disable the editor's actions to avoid conflicts. For this,
add a boolean property to the strategy API.

Implement a new editor size handler for the reworked
terminal. For that, provide the font size service
as a data key in the editor, so the actions
located in execution-impl can access the service
internal to the terminal otherwise.

While we're at it, introduce action text overrides for
Search Everywhere to avoid confusion between editor's
and terminal's actions.

Note that the reset font size actions are available
from both the terminal and the editor. This is,
unfortunately, necessary at this point because
that's how that "Reset to XXpt" link works at the moment.
This can be fixed in the future if necessary.

GitOrigin-RevId: 7c96b64ed1b7d7330a2c64198fb49a45820c2912