[Kotlin] Report recursive property access warning only for the same receiver
Recursive property call on the same receiver instance is a likely error that
can lead to infinite recursion. A call for the same property but with a
different receiver can be intentional. A gutter icon for a recursive
call should be enough in such cases.

Additionally, the quick fix that replaces a recursive call with a `field` call
is no longer suggested for properties without a backing field.

KTIJ-24961

GitOrigin-RevId: 55fb4f25063dbe63375f7bf5c22b1ef0b1179df1