IJPL-184027 Fix font family saving/restoring in TerminalFontOptions
The whole machinery of font options, preferences, persistent preferences
etc. is build on the idea of having a font family and two sub-families:
the regular and the bold. Failure to specify them can lead to different
results. The default state constructor, for instance, does not default
to null values, but takes some default values from a service.
But when copying preferences to persistent preferences,
then values are copied as-is, including nulls. This means that a null
value is not the default value.

Then these null values are treated specially when loading the state,
and some migration mechanism kicks in in
com.intellij.openapi.editor.colors.impl.AppFontOptions.copyState
that for some reason corrupts the main font family in the case
of this issue. Why that happens is another story,
but common sense tells us that no migration should happen
here to begin with, as we're dealing with a brand new feature here.

Therefore, fix by actually storing values retrieved from the same
service that is used in the default PersistentFontPreferences
constructor.

GitOrigin-RevId: f7250ed546739c01441adf1f564084ea8abe0613