IJPL-184236 Introduce EditorColorsManager.getDefaultScheme
The problem is that getScheme() is nullable, even though
it's not annotated as such. We want to annotate it,
but it'll break every getScheme(getDefaultSchemeName())
call, and there's a lot of them. So we first provide
a default hack to keep the code compiling
and then do the essential part in a separate commit
that will be much smaller.

GitOrigin-RevId: 39b35af14fca1d5472a85ee6962ef23a57813545