IJPL-184236 Hotfix for NPE in LafManagerImpl.syncEditorScheme
The core reason for the NPE is EditorColorsManager.getScheme
is not annotated as neither nullable nor non-nullable.
This lead to a subtle bug with Kotlin using the "not really sure"
nullability type, and, as a result, missing null check.

It's not really clear what we should do in case of a null
there. Perhaps some fallback is needed for a proper fix.
But for a hotfix, we can at least do nothing, which is
way better than an NPE that prevents the IDE startup at that.

Now that we got rid of most getScheme() calls
by replacing them with the new getDefaultScheme(),
there are these places that need fixing:

- the hotfix itself in LafManagerImpl.syncEditorScheme;
- FilenameToolbarWidgetAction.updatePresentationFromFile,
added a null-safe call there, as there's already a fallback for null.

GitOrigin-RevId: f1e8b9e7269e0c3dda42780dcc8c6d0048b1c171