[debugger] Collect ALL continuations up to the CoroutineOwner for the top stack frame.
`CoroutinesDebugHelper#getCoroutineStackTraceDump` added a frame to the async stack trace only if the corresponding continuation was an instance of `kotlin.coroutines.BaseContinuationImpl`. But that's not always true, e.g. for `DispatchedContinuation` in case of `withContext`.

So, collect frames with `kotlin.coroutines.CoroutineStackFrame#callerFrame` and extract spilled variables if a continuation is an instance of `BaseContinuationImpl`

IDEA-370748

GitOrigin-RevId: 4a722eba06bbc6e893d9f4edba3ac773510c7ea0