[terminal] IJPL-183992 Pass startup FUS info to the ReworkedTerminalWidget
The single reliable way is to pass an object with metadata about terminal startup through the process of creating the terminal widget to later use it for reporting.
Mark all related exposed APIs as internal.
We will report the metrics only if the terminal was opened from the known places to be sure about the execution flow.

(cherry picked from commit bba153cb4637510bcf15385307103031b186b36e)

IJ-CR-160917

GitOrigin-RevId: 4c8aad3efd82aea669865e3db9c6b5253616a0a1