IJPL-184686 Add TreeBaseConstructorSafetyTest
This will contain various tests that ensure that a tree
can be safely constructed regardless of what overridden methods
are called from base class constructors, directly or indirectly.

For now, just a very basic test and a simple fix that ensures
that the class can be used in unit tests without an Application instance.

GitOrigin-RevId: 75e0c083ce38d60cf479205b272c6b1d37399361