IJPL-184686 Support collapsing the root from Tree base constructors
The use case: some class overrides setModel() and calls
setExpandedState(rootPath, false) there to override
the default JTree logic of auto-expanding the root.

Fix by relying on getExpandedDescendants exclusively
in ExpandImpl() instead of trying to mimic what setModel
does, because we don't know what it does in case
it was overridden.

GitOrigin-RevId: 17361bd49d0a1e8c65e561c4b7714a91600838be