[driver] Use write-intent read action as the read lock for EDT computations
By default, NO_LOCK semantics leads to the complete absence of lock. We need to execute some runnables under write-intent lock, and we don't want to pollute the API layer by an internal concept. Hence, the combination of OnDispatcher.EDT and LockSemantics.READ_ACTION will lead to the Write-Intent lock on EDT, as a way to perform computations in compatibility mode. We hope that no one wants to run pure read computations on EDT

GitOrigin-RevId: 8ae0706f0aba2f3e6cbef42c6c45b361097f7c19