[git] IJPL-173922 Replace rhizome with rpc
The reason is the lack of the update visibility guarantees.
E.g., obsolete state can be returned when accessing entity from EDT after the change was applied.
Switching to rpc we expect more control on what, when and how is synced. However, it requires some extra implementation efforts, which might be compensated by the elimination of rhizome obscurity.

GitOrigin-RevId: ff271274716d164e270f4a849760c92d12ccb427