IJPL-180915 gitlab: do not specify `<idea-plugin>`'s `package` attribute in module xml-s
also do not specify `allow-bundled-update` attribute, because it makes no sense in the xml belonging to a module.

The moving of `GitlabIcons.java` and the changes in `plugin-content.yaml` is needed for fixing tests `CommunityIconClassesTest.test` and `IdeaUltimatePackagingTest.test`.

`<idea-plugin>`'s `package` attribute should be specified in a `plugin.xml` file of a corresponding plugin. It shouldn't be specified in the xml-file corresponding to any module which is belonged to the plugin.

Specifying `package` in the xml-file corresponding to a module can lead to an exception like

```
class XXX cannot be cast to class XXX (XXX is in unnamed module of loader ...PluginClassLoader @YYY; cXXX is in unnamed module of loader ...PluginClassLoader @ZZZ)
```

To get this error create 2 modules where "package" attributes will be specified in both corresponding xml files. Write a service (https://plugins.jetbrains.com/docs/intellij/plugin-services.html) `XXX` in any module. Write the code so that in runtime the service is called first from the code located in one module, and later the service is called first from the code located in the second module. On the latter call the exception will appear.

The problem appears because of some magic interaction between class loaders and our plugins-modules system.

The changes were authored by Nikita Katkov.
The commit message was written by Vladimir Petrenko.
Commited by Vladimir Petrenko.
It's better to direct any related questions to Nikita Katkov.

GitOrigin-RevId: 33d46eb7056fed39ed4e3ad9d19578f9befd0437