IJPL-184075 Validate the terminal editor size when changing the font
The standard editor implementation doesn't work because
it has a bug: it doesn't update the size immediately,
because it calls invalidate() and not validateSize().
This is not a big deal for the editor, as it almost never
uses a point that's close to the bottom,
and it's actually beneficial in terms of performance.
But it's a deal breaker for the terminal,
as increasing the font size never scrolls correctly.
Fix by introducing a flag therefore.

(cherry picked from commit 9b5ba77a5aa314b9ca8d3a997b5705fc54b7bea7)

IJ-CR-160631

GitOrigin-RevId: 1c58b1e61e464a13a7558e26ca4d0a25e09156d5