[gh/gl] Explicitly provide editor tab titles (IJPL-183254)
#IJPL-183254 Fixed

The editor tab title-providing mechanism is a little wonky.
It relies on some check to see if the `getPath` method for Virtual files contains something that looks realish.
The one for GitHub turns out to look realish, but the one for GitLab does not.
Better not to rely on that and use the explicit mechanism.

See: com.intellij.openapi.wm.impl.PlatformFrameTitleBuilderKt.doGetFileTitle

GitOrigin-RevId: 24483266aa97cf81a1b0486eb41949681fe6fc49