[Kotlin] Literal copy-paste: improve pasting from external sources
When pasting into multi-quoted strings from outside the IDE,
the overall logic now is to treat everything as plain text.
This means that pasted characters that need escaping will be escaped.
The interpolation prefix of the host string is considered
so that only unsafe characters are escaped.

For comparison, previously if something could be interpreted as an
interpolation entry, it was not escaped.

The commit also removes the adhoc escaping for stacktraces as it is
now covered by the main logic.

KTIJ-33667

GitOrigin-RevId: 96c696cf988ed7046f9d315d9ac7edc0d6fd7fcd