[rdct] IJPL-182411: Do not focus on Project View on RD's backend
Since the current implementation of Structure View is LUXed, the focus on the backend should be precisely set to gather the correct PSI information. That includes, for example, that the focus for the editor is not set correctly on the backend. So, if on the startup the focus is set to Project View, the Structure View will be empty until some LUXed window change the focus

GitOrigin-RevId: c4a46cfbfceeda4178c32e34f9974ab871de9af0