IJPL-185303 Do not spam UiNotifyConnector instances in toolbars
When a toolbar is added/removed many times to a parent
that is not currently showing, on every addNotify it would
add a listener through UiNotifyConnector. Eventually there
can be too many such listeners to handle.

Fix by using launchOnceOnShow instead and saving
the returned job. If there's already a job scheduled,
simply do nothing. Remove the job to save memory
when we're done. This, of course, has a side effect
of launching the same job again if the toolbar is
removed and then added. That is not an issue because
updateActionsFirstTime() is a no-op unless
it's a really first update, so subsequent jobs
won't do anything. And because there's at most
one job, it'll never become a performance problem.

Dispatchers.EDT is needed in launchOnceOnShow
because updateActionsImmediately() is a legacy
API that requires both EDT and read actions.
By default launchOnceOnShow uses Dispatchers.UI
that prohibits read actions.

GitOrigin-RevId: 97f77a0d8192a0457c1a7e1e9395c21534705800