IJ-CR-160631 Make TerminalFontOptions precisions consistent
We have three classes that had an inner precision of 0.0001,
but in the UI they were displayed with the precision of 0.1.
Meaning visually identical values were considered different.

Fix by introducing explicit precision and encapsulating
formatting and parsing. For the font size leave it as 0.1,
but for spacings let's increase it to 0.01, as it seems
useful.

(cherry picked from commit 837b11582f1985416af42eaa174135f7c235c412)

IJ-CR-160631

GitOrigin-RevId: 33fd47705baf53cbf2eebcf864f11b990fa76d1e