Do not log an error when retrieving Gradle JVM path
This is a short term fix to no longer have logger.errors thrown when
accessing the Gradle Jvm Path on the EDT, or without a progress
indicator. These are known issues introduced in the 2025.1 merge which
will require some more time and collaboration to address upstream.

Two errors impacting a lot of our tests:
1. "Caused by: java.lang.IllegalStateException: This
method is forbidden on EDT because it does not pump the event queue.
Switch to a BGT, or use "
2. "There is no ProgressIndicator or Job in this thread, the current job is
not cancellable."

Upstream issue: https://youtrack.jetbrains.com/issue/IDEA-370663

Bug: 411744564
Change-Id: I8e58725577cd3ee68bae424d7d304207be751fdf