[fleet] do not use hackyNonBlockingChange for durable snapshots
it is possible for frontend to have an entity type that is not accessible from the workspace layer, but is stored in the shared partition with the workspace storage key

the idea is that workspace will populate the shared partition with the snapshot contents using raw json objects for values and then frontend will reify these entities and deserialize all the attributes. deserialization happens either when the transaction is applied if we already have a serializer, or on plugin loading

when running in short circuit mode, hackyNonBlockingChange might
capture a version of the database that does not contain an entity type
with its serializers, but will be applied to a version that already
has it. the change will be expanded into a plain assertion of json
values that will never be deserialized later

FL-33229 "Open" doesn't work from the welcome screen

GitOrigin-RevId: d5f87eab099d2be273f203f8cf43d3a9ab8c6018