[rd][debugger] Don't create new FrontendXSuspendContext on pause if it didn't change in the backend
Implementations of XDebugger API are free to call pause on any convenient occasion. In Java, Async Stack Traces toggle did that, unintentionally, until recently. And it basically means that all context, stacks and frames might not change on pause. But stacks and frames are, unlike suspend context, deduplicated on the frontend side, and thus they outlive the suspend context, and this is undesired behavior.

GitOrigin-RevId: 76d4cae4016283dff124ffcd9996288b883e6a30