[terminal] IJPL-184902 Open the selected shell from the predefined sessions list instead of default (classic terminal)
To make it possible, we need to propagate the shell command (used for starting the process) from the frontend to the backend. This shell command is preset when you select some predefined session.
Create the separate field in the `RdTerminalSession` model for that. This field will be used only when the session was created on the frontend. On the backend we will just init it as null and it will be unused.

GitOrigin-RevId: e675ba4dad0c791cda67d5c78b7968ed93a609cf